feat(detect): Suppression des anciens fichiers + ajout de la direction sur l'affichage + ajout de la possibilité de ne pas enregistrer les données dans le CSV
