--- Functions/sort.py
@@ -1,254 +1,0 @@
-"""
-    https://github.com/abewley/sort
-
-    SORT: A Simple, Online and Realtime Tracker
-    Copyright (C) 2016-2020 Alex Bewley alex@bewley.ai
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-"""
-
-import numpy as np
-import torch
-from filterpy.kalman import KalmanFilter
-
-np.random.seed(0)
-
-
-def iou_batch(bb_test: torch.Tensor, bb_gt: np.ndarray) -> np.ndarray:
-    """
-    From SORT: Computes IUO between two bboxes in the form [x1,y1,x2,y2]
-    @param bb_test: (N,4) ndarray of float
-    @param bb_gt: (K,4) ndarray of float
-    @return: ndarray of overlap between boxes and query_boxes
-    """
-    bb_gt = np.expand_dims(bb_gt, 0)
-    bb_test = np.expand_dims(bb_test, 1)
-
-    xx1 = np.maximum(bb_test[..., 0], bb_gt[..., 0])
-    yy1 = np.maximum(bb_test[..., 1], bb_gt[..., 1])
-    xx2 = np.minimum(bb_test[..., 2], bb_gt[..., 2])
-    yy2 = np.minimum(bb_test[..., 3], bb_gt[..., 3])
-    w = np.maximum(0., xx2 - xx1)
-    h = np.maximum(0., yy2 - yy1)
-    wh = w * h
-    o = wh / ((bb_test[..., 2] - bb_test[..., 0]) * (bb_test[..., 3] - bb_test[..., 1])
-              + (bb_gt[..., 2] - bb_gt[..., 0]) * (bb_gt[..., 3] - bb_gt[..., 1]) - wh)
-    return o
-
-
-def convert_bbox_to_z(bbox: torch.Tensor) -> np.ndarray:
-    """
-    Takes a bounding box in the form [x1,y1,x2,y2] and returns z in the form
-      [x,y,s,r] where x,y is the centre of the box and s is the scale/area and r is
-      the aspect ratio
-    @param bbox: (N,4) ndarray of float
-    @return: (N,5) ndarray of float
-    """
-    w = bbox[2] - bbox[0]
-    h = bbox[3] - bbox[1]
-    x = bbox[0] + w / 2.
-    y = bbox[1] + h / 2.
-    s = w * h  # scale is just area
-    r = w / float(h)
-    return np.array([x, y, s, r]).reshape((4, 1))
-
-
-def convert_x_to_bbox(x: np.ndarray) -> np.ndarray:
-    """
-    Takes a bounding box in the centre form [x,y,s,r] and returns it in the form
-      [x1,y1,x2,y2] where x1,y1 is the top left and x2,y2 is the bottom right
-    @param x: (N,5) ndarray of float
-    @return: (N,4) ndarray of float
-    """
-    w = np.sqrt(x[2] * x[3])
-    h = x[2] / w
-    return np.array([x[0] - w / 2., x[1] - h / 2., x[0] + w / 2., x[1] + h / 2.]).reshape((1, 4))
-
-
-class KalmanBoxTracker(object):
-    """
-    This class represents the internal state of individual tracked objects observed as bbox.
-    """
-    count = 0
-
-    def __init__(self, params: torch.Tensor) -> None:
-        """
-        Initialises a tracker using initial bounding box.
-        @param params: (N,6) ndarray of float
-        """
-        # define constant velocity model
-        self.kf = KalmanFilter(dim_x=7, dim_z=4)
-        self.kf.F = np.array(
-            [[1, 0, 0, 0, 1, 0, 0], [0, 1, 0, 0, 0, 1, 0], [0, 0, 1, 0, 0, 0, 1], [0, 0, 0, 1, 0, 0, 0],
-             [0, 0, 0, 0, 1, 0, 0], [0, 0, 0, 0, 0, 1, 0], [0, 0, 0, 0, 0, 0, 1]])
-        self.kf.H = np.array(
-            [[1, 0, 0, 0, 0, 0, 0], [0, 1, 0, 0, 0, 0, 0], [0, 0, 1, 0, 0, 0, 0], [0, 0, 0, 1, 0, 0, 0]])
-
-        self.kf.R[2:, 2:] *= 10.
-        self.kf.P[4:, 4:] *= 1000.  # give high uncertainty to the unobservable initial velocities
-        self.kf.P *= 10.
-        self.kf.Q[-1, -1] *= 0.01
-        self.kf.Q[4:, 4:] *= 0.01
-
-        self.kf.x[:4] = convert_bbox_to_z(params)
-        self.time_since_update = 0
-        self.id = KalmanBoxTracker.count
-        KalmanBoxTracker.count += 1
-        self.history = []
-        self.hits = 0
-        self.hit_streak = 0
-        self.age = 0
-        self.obj_confidence = params[4]
-        self.obj_class = params[5]
-
-    def update(self, params: torch.Tensor) -> None:
-        """
-        Updates the state vector with observed bbox.
-        @param params: (N,6) ndarray of float
-        """
-        self.time_since_update = 0
-        self.history = []
-        self.hits += 1
-        self.hit_streak += 1
-        self.kf.update(convert_bbox_to_z(params))
-        self.obj_confidence = params[4]
-        self.obj_class = params[5]
-
-    def predict(self) -> np.ndarray:
-        """
-        Advances the state vector and returns the predicted bounding box estimate.
-        @return: (N,4) ndarray of float
-        """
-        if (self.kf.x[6] + self.kf.x[2]) <= 0:
-            self.kf.x[6] *= 0.0
-        self.kf.predict()
-        self.age += 1
-        if self.time_since_update > 0:
-            self.hit_streak = 0
-        self.time_since_update += 1
-        self.history.append(convert_x_to_bbox(self.kf.x))
-        return self.history[-1]
-
-    def get_state(self) -> np.ndarray:
-        """
-        Returns the current bounding box estimate.
-        @return: (N,4) ndarray of float
-        """
-        return convert_x_to_bbox(self.kf.x)
-
-
-def associate_detections_to_trackers(detections: torch.Tensor, trackers: np.ndarray, iou_threshold: float = 0.3) -> \
-        tuple[np.ndarray, np.ndarray, np.ndarray]:
-    """
-    Assigne les détections aux objets suivis (les deux représentés sous forme de boîtes englobantes)
-    @param detections:liste des détections
-    @param trackers: liste des objets suivis
-    @param iou_threshold: Seuil de similarité IOU pour l'association entre les détections et les trackers
-    @return: Les trackers valides, les détections non associées et les trackers morts
-    """
-    if len(trackers) == 0:
-        return np.empty((0, 2), dtype=int), np.arange(len(detections)), np.empty((0, 5), dtype=int)
-
-    iou_matrix = iou_batch(detections, trackers)
-
-    if min(iou_matrix.shape) > 0:
-        a = (iou_matrix > iou_threshold).astype(np.int32)
-        matched_indices = np.stack(np.where(a), axis=1)
-    else:
-        matched_indices = np.empty(shape=(0, 2))
-
-    unmatched_detections = []
-    for d, _ in enumerate(detections):
-        if d not in matched_indices[:, 0]:
-            unmatched_detections.append(d)
-    unmatched_trackers = []
-    for t, _ in enumerate(trackers):
-        if t not in matched_indices[:, 1]:
-            unmatched_trackers.append(t)
-
-    # filtrer les correspondances avec un faible IOU
-    matches = []
-    for m in matched_indices:
-        if iou_matrix[m[0], m[1]] < iou_threshold:
-            unmatched_detections.append(m[0])
-            unmatched_trackers.append(m[1])
-        else:
-            matches.append(m.reshape(1, 2))
-    if len(matches) == 0:
-        matches = np.empty((0, 2), dtype=int)
-    else:
-        matches = np.concatenate(matches, axis=0)
-
-    return matches, np.array(unmatched_detections), np.array(unmatched_trackers)
-
-
-class Sort(object):
-    def __init__(self, max_age: int = 1, min_hits: int = 3, iou_threshold: float = 0.3) -> None:
-        """
-        Initialise les paramètres clés pour SORT
-        @param max_age: Durée maximale d'un tracker avant qu'il soit considéré comme mort
-        @param min_hits: Nombre minimum de fois où un tracker doit être associé à une détection
-         pour être considéré comme valide
-        @param iou_threshold: Seuil de similarité IOU pour l'association entre les détections et les trackers
-        """
-        self.max_age = max_age
-        self.min_hits = min_hits
-        self.iou_threshold = iou_threshold
-        self.trackers = []
-        self.frame_count = 0
-
-    def update(self, dets: torch.Tensor = np.empty((0, 6))) -> np.ndarray:
-        """
-        Met à jour les trackers en utilisant les détections fournies
-        Note : Le nombre d'objets retournés peut être différent du nombre de détections fournies.
-        @param dets: Tableau des détections au format [[x1, y1, x2, y2, confidence, classe], ...]
-        @return: Même tableau que les détections, avec en plus l'ID du tracker associé à la position 4
-        """
-        self.frame_count += 1
-        # obtenir les emplacements prévus à partir des trackers existants
-        trks = np.zeros((len(self.trackers), 5))
-        to_del = []
-        ret = []
-        for t, trk in enumerate(trks):
-            pos = self.trackers[t].predict()[0]
-            trk[:] = [pos[0], pos[1], pos[2], pos[3], 0]
-            if np.any(np.isnan(pos)):
-                to_del.append(t)
-        trks = np.ma.compress_rows(np.ma.masked_invalid(trks))
-        for t in reversed(to_del):
-            self.trackers.pop(t)
-        matched, unmatched_dets, _ = associate_detections_to_trackers(dets, trks, self.iou_threshold)
-
-        # mettre à jour les trackers associés avec les détections correspondantes
-        for m in matched:
-            self.trackers[m[1]].update(dets[m[0], :])
-
-        # créer et initialiser de nouveaux trackers pour les détections non associées
-        for i in unmatched_dets:
-            trk = KalmanBoxTracker(dets[i, :])
-            self.trackers.append(trk)
-        i = len(self.trackers)
-        for trk in reversed(self.trackers):
-            d = trk.get_state()[0]
-            if (trk.time_since_update < 1) and (trk.hit_streak >= self.min_hits or self.frame_count <= self.min_hits):
-                ret.append(np.concatenate((d, [trk.id + 1], [trk.obj_confidence], [trk.obj_class])).reshape(1, -1))
-                # +1, car MOT benchmark nécessite des ID positifs
-            i -= 1
-            # supprimer les trackers morts
-            if trk.time_since_update > self.max_age:
-                self.trackers.pop(i)
-        if len(ret) > 0:
-            return np.concatenate(ret)
-        return np.empty((0, 6))
