+++ config.ini
@@ -11,6 +11,8 @@ display_detection = False
 display_fps = True
 # La valeur par défaut est `False`, si vous voulez activer l'affichage des messages, mettez `True`
 debug = False
+# La valeur par défaut est `True`, si vous voulez desactiver l'enregistrement dans un fichier csv, mettez `False`
+save_in_csv = True
 
 [YOLOV5_PARAMS]
 # La valeur par défaut est `yolov5n`
